pyproject.toml
[project]
name = "teachio"
version = "0.2"
description = "Telegram bot management teacher's tasks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "click>=8.3.1",
    "dotenv==0.9.9",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "python-dotenv==1.1.1",
    "python-telegram-bot==22.4",
    "sniffio==1.3.1",
    "toml>=0.10.2",
    "watchdog==6.0.0",
]
